refactor: consolidate components
